r562
Удаление мода виртуальной клавиатуры из основного дистрибутива; фикс главной страницы админ-панели; фикс панели управления шаблонами в админ-панели; частичная замена else if на elseif, для возможности отхода от использования фигурных скобок в дальнейшем. Дальнейшая поддержка и распространение мода виртуальной клавиатуры осуществляется на форуме: http://torrentpier.me/resources/84/ git-svn-id: https://torrentpier2.googlecode.com/svn/trunk@562 a8ac35ab-4ca4-ca47-4c2d-a49a94f06293

+++ upload/admin/admin_disallow.php
@@ -41,12 +41,11 @@ if( isset($_POST['add_name']) )
 
 	message_die(GENERAL_MESSAGE, $message);
 }
-else if( isset($_POST['delete_name']) )
+elseif (isset($_POST['delete_name']))
 {
-	$disallowed_id = ( isset($_POST['disallowed_id']) ) ? intval( $_POST['disallowed_id'] ) : intval( $_GET['disallowed_id'] );
+	$disallowed_id = (isset($_POST['disallowed_id']) ) ? intval( $_POST['disallowed_id'] ) : intval( $_GET['disallowed_id']);
 
-	$sql = "DELETE FROM " . BB_DISALLOW . "
-		WHERE disallow_id = $disallowed_id";
+	$sql = "DELETE FROM " . BB_DISALLOW . " WHERE disallow_id = $disallowed_id";
 	$result = DB()->sql_query($sql);
 	if( !$result )
 	{
